+++ NEWS.md
@@ -1,6 +1,7 @@
 ## version 0.2.16
 ### Features
 - update an existing JAGS fit with `JAGS_extend()` function
+- new element of the `autofit_control` argument in `JAGS_fit()`: `"restarts"` allows to restart model initialization up to `restarts` times in case of failure
 
 ## version 0.2.15
 ### Fixes